Finish hero section
